+++ core/src/main/java/arez/component/internal/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains internal implementation classes required to support the component model used by Arez.
+ */
+package arez.component.internal;