+++ core/src/test/java/arez/component/internal/MemoizeCacheTest.java
@@ -1,4 +1,4 @@
-package arez.component;
+package arez.component.internal;
 
 import arez.AbstractArezTest;
 import arez.Arez;