+++ processor/src/test/resources/expected/com/example/id/Arez_RepositoryExample.java
@@ -4,11 +4,11 @@ import arez.Arez;
 import arez.ArezContext;
 import arez.Component;
 import arez.Disposable;
-import arez.component.ComponentKernel;
 import arez.component.ComponentObservable;
 import arez.component.DisposeNotifier;
 import arez.component.DisposeTrackable;
 import arez.component.Identifiable;
+import arez.component.internal.ComponentKernel;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;