+++ processor/src/test/resources/expected/com/example/repository/InjectEnabledRepositoryRepository.java
@@ -3,7 +3,7 @@ package com.example.repository;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.Feature;
-import arez.component.AbstractRepository;
+import arez.component.internal.AbstractRepository;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.inject.Singleton;